awssam/devfile/requirements.txt

pyflakes==2.0.0
Pygments==2.2.0
PyHive==0.6.1
PyJWT==1.7.1
pylint==2.1.1
pyneuroner==1.0.6
pyodbc==4.0.24
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
pyspark==2.3.2
pytest==3.8.0
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-bioformats==1.5.2
python-dateutil==2.7.3
python-dotenv==0.10.1
python-jsonrpc-server==0.1.2
python-language-server==0.26.1
python3-openid==3.1.0
pytz==2018.5
PyWavelets==1.0.0
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.4.1
QtPy==1.5.0
regex==2019.4.14
requests==2.19.1
requests-oauthlib==1.2.0
rope==0.11.0
ruamel-yaml==0.15.46
s3transfer==0.2.0
scikit-image==0.14.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
seqeval==0.0.10
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.8.1
snowballstemmer==1.2.1
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sortedcollections==1.0.1
sortedcontainers==2.0.5
spacy==2.1.3
Sphinx==1.7.9
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
spylon==0.3.0
spylon-kernel==0.4.1
SQLAlchemy==1.2.11
sqlparse==0.2.4
srsly==0.0.5
stanford-corenlp==3.9.2
statsmodels==0.9.0
sympy==1.2
tables==3.4.4
tblib==1.3.2
tensorflow==0.12.0
terminado==0.8.1
testpath==0.3.1
text-unidecode==1.2
thinc==7.0.4
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tornado==5.1
tqdm==4.26.0
traitlets==4.3.2
Twisted==18.7.0
unicodecsv==0.14.1
urllib3==1.23
wasabi==0.2.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==4.1.2
widgetsnbextension==3.4.1
wrapt==1.10.11
xgboost==0.80
xlrd==1.1.0
XlsxWriter==1.1.0
xlwings==0.11.8
xlwt==1.3.0
xmltodict==0.12.0
xmlutils==1.4
yellowbrick==0.9
zict==0.1.3
zope.interface==4.5.0
